Set up CSS variables/custom properties
The setup for the colors is a bit involved due to the different areas
in the design where different opacity levels are needed.
The font weights are provided in the design.
The font sizes are simply guesses as the design provided is a JPEG
from which the font sizes cannot be exactly determined.