Add general styling to results section
Background color is set.
Typography properties, such as text alignment and text colors,
are set on the parent element (.results) so they are inherited
by all the children.
Background radius is set on the top-level .results-summary element
so that it is inherited by all its descendants, particularly
.results and .summary elements.
CSS property "overflow: hidden" is used to hide the overflow
and reveal the border radius around the edges of .results-summary element.