Fix issue with border radius in responsive design
The border radii at the top left and right corners also show up in
the responsive design when they should not.
A bit of margin is also added on the left and right sides of the component
so that there is space between the edge of the screen and the component.