Set up layout for summary items
Flexbox is used for the layout within each summary item.
A background color of light blue and a stroke of red are temporarily
used to display each item and SVG respectively so that they are visible
for easier layout and styling. These will be changed to the
correct values later.