Style summary items
Colors are added to the SVG, text and background of the summary items.
A different color is applied to each item by declaring a custom property
that sets the color in the container element, set each to a different color,
and then letting it get inherited by the elements within the container.